Create asset pipeline framework pattern
Most other engines depend on the nodejs engine to provide a javascript
asset compilation process if the application needs it. Currently, there
are two issues with the implementation:

1) It's explicit and tedious for the other non-nodejs engines and
2) it assumes bower is the only package manager in existence.

This commit establishes a clear pattern for javascript asset toolchains
to provide a modular implementation. Each plugin will have to implement
a series of functions to work.